+++ core/app/c/[communitySlug]/types/[pubTypeId]/edit/FieldBlock.tsx
@@ -2,7 +2,7 @@ import { useId } from "react";
 import { useSortable } from "@dnd-kit/sortable";
 import { CSS } from "@dnd-kit/utilities";
 import { type Static } from "@sinclair/typebox";
-import { ArchiveRestore, GripVertical, Pencil, Text, Trash, TypeOutline } from "lucide-react";
+import { ArchiveRestore, GripVertical, Trash, TypeOutline } from "lucide-react";
 
 import { Button } from "ui/button";
 import { Tooltip, TooltipContent, TooltipTrigger } from "ui/tooltip";
@@ -10,6 +10,7 @@ import { cn } from "utils";
 
 import { useBuilder } from "~/app/components/FormBuilder/BuilderContext";
 import { FieldIcon } from "~/app/components/FormBuilder/FieldIcon";
+import { pubFieldCanBeTitle } from "../../utils";
 import { pubTypeBuilderSchema } from "./TypeBuilder";
 
 export type FieldThingProps = {
@@ -21,7 +22,7 @@ export type FieldThingProps = {
 	toggleTitle: () => void;
 };
 
-export const FieldThing = ({
+export const FieldBlock = ({
 	field,
 	index,
 	isEditing,
@@ -116,34 +117,33 @@ export const FieldThing = ({
 						</div>
 					</div>
 				</>
-				{/* {isEditing ? (
-					<div className="my-auto ml-auto text-xs text-blue-500">EDITING</div>
-				) : ( */}
 				<div className="my-auto ml-auto flex gap-1">
-					<Tooltip delayDuration={300}>
-						<TooltipTrigger asChild>
-							<Button
-								type="button"
-								aria-label="Set as title"
-								disabled={isDisabled || field.deleted}
-								variant="ghost"
-								className={cn(
-									"p-1.5 text-neutral-400 opacity-0 group-focus-within:opacity-100 group-hover:opacity-100",
-									"hover:bg-white hover:text-blue-500",
-									isTitle && "text-blue-500 opacity-100"
-								)}
-								onClick={() => {
-									toggleTitle();
-								}}
-								tabIndex={0}
-							>
-								<TypeOutline size={24} />
-							</Button>
-						</TooltipTrigger>
-						<TooltipContent>
-							Mark this field as the title of this Pub Type
-						</TooltipContent>
-					</Tooltip>
+					{pubFieldCanBeTitle(field) && (
+						<Tooltip delayDuration={300}>
+							<TooltipTrigger asChild>
+								<Button
+									type="button"
+									aria-label="Set as title"
+									disabled={isDisabled || field.deleted}
+									variant="ghost"
+									className={cn(
+										"p-1.5 text-neutral-400 opacity-0 group-focus-within:opacity-100 group-hover:opacity-100",
+										"hover:bg-white hover:text-blue-500",
+										isTitle && "text-blue-500 opacity-100"
+									)}
+									onClick={() => {
+										toggleTitle();
+									}}
+									tabIndex={0}
+								>
+									<TypeOutline size={24} />
+								</Button>
+							</TooltipTrigger>
+							<TooltipContent>
+								Mark this field as the title of this Pub Type
+							</TooltipContent>
+						</Tooltip>
+					)}
 
 					{restoreRemoveButton}
 					{/* Cant edit fields yet */}
