+++ core/app/c/[communitySlug]/types/actions.ts
@@ -1,8 +1,6 @@
 "use server";
 
-import { sql } from "kysely";
-
-import type { CommunitiesId, PubFieldsId, PubTypesId } from "db/public";
+import type { CommunitiesId, PubFieldsId } from "db/public";
 import { Capabilities, MembershipType } from "db/public";
 import { logger } from "logger";
 
@@ -21,223 +19,6 @@ import {
 	FORM_SLUG_UNIQUE_CONSTRAINT,
 	insertForm,
 } from "~/lib/server/form";
-
-export const addPubField = defineServerAction(async function addPubField(
-	pubTypeId: PubTypesId,
-	pubFieldId: PubFieldsId
-) {
-	const [loginData, community] = await Promise.all([getLoginData(), findCommunityBySlug()]);
-	if (!loginData || !loginData.user) {
-		return ApiError.NOT_LOGGED_IN;
-	}
-
-	if (!community) {
-		return ApiError.COMMUNITY_NOT_FOUND;
-	}
-
-	const { user } = loginData;
-
-	const [authorized, pubType] = await Promise.all([
-		await userCan(
-			Capabilities.editPubType,
-			{ type: MembershipType.community, communityId: community.id },
-			user.id
-		),
-		getPubType(pubTypeId).executeTakeFirst(),
-	]);
-
-	if (!authorized) {
-		return ApiError.UNAUTHORIZED;
-	}
-
-	if (!pubType) {
-		return ApiError.PUB_TYPE_NOT_FOUND;
-	}
-
-	const newRank = findRanksBetween({
-		numberOfRanks: 1,
-		start: pubType.fields.at(-1)?.rank ?? "0",
-	});
-
-	await autoRevalidate(
-		db.insertInto("_PubFieldToPubType").values({
-			A: pubFieldId,
-			B: pubTypeId,
-			rank: newRank[0],
-		})
-	).execute();
-});
-
-export const updatePubType = defineServerAction(async function updatePubType(opts: {
-	pubTypeId: PubTypesId;
-	name?: string;
-	description?: string | undefined;
-	fields: {
-		id: PubFieldsId;
-		rank: string;
-		deleted: boolean;
-		isTitle: boolean;
-	}[];
-}) {
-	const [{ user }, community] = await Promise.all([getLoginData(), findCommunityBySlug()]);
-
-	if (!user) {
-		return ApiError.NOT_LOGGED_IN;
-	}
-
-	if (!community) {
-		return ApiError.COMMUNITY_NOT_FOUND;
-	}
-
-	const authorized = await userCan(
-		Capabilities.editPubType,
-		{ type: MembershipType.community, communityId: community.id },
-		user.id
-	);
-
-	if (!authorized) {
-		return ApiError.UNAUTHORIZED;
-	}
-
-	const { pubTypeId, name, description, fields } = opts;
-
-	const fieldsToDelete = fields.filter((field) => field.deleted);
-
-	const fieldsToUpsert = fields
-		.filter((field) => !field.deleted)
-		.map((field) => ({
-			...field,
-			isTitle: field.isTitle,
-		}));
-
-	try {
-		const res = await db.transaction().execute(async (trx) => {
-			if (name || description !== undefined) {
-				await autoRevalidate(
-					trx
-						.updateTable("pub_types")
-						.set({ name, description })
-						.where("id", "=", pubTypeId)
-				).execute();
-			}
-
-			if (fieldsToDelete.length > 0) {
-				await autoRevalidate(
-					trx
-						.deleteFrom("_PubFieldToPubType")
-						.where("B", "=", pubTypeId)
-						.where(
-							"A",
-							"in",
-							fieldsToDelete.map((field) => field.id)
-						)
-				).execute();
-			}
-
-			if (fieldsToUpsert.length > 0) {
-				await autoRevalidate(
-					trx
-						.insertInto("_PubFieldToPubType")
-						.values(
-							fieldsToUpsert.map((field) => ({
-								A: field.id,
-								B: pubTypeId,
-								isTitle: field.isTitle,
-								rank: field.rank,
-							}))
-						)
-						.onConflict((b) =>
-							b.columns(["A", "B"]).doUpdateSet((eb) => ({
-								isTitle: sql<boolean>`excluded."isTitle"`,
-								rank: sql<string>`excluded.rank`,
-							}))
-						)
-				).execute();
-			}
-		});
-
-		return {
-			data: {},
-			success: true,
-		};
-	} catch (error) {
-		logger.error(error);
-		return { error: "Pub type update failed", cause: error };
-	}
-});
-
-export const updateTitleField = defineServerAction(async function updateTitleField(
-	pubTypeId: PubTypesId,
-	pubFieldId: PubFieldsId
-) {
-	const loginData = await getLoginData();
-	if (!loginData || !loginData.user) {
-		return ApiError.NOT_LOGGED_IN;
-	}
-
-	const { user } = loginData;
-
-	const community = await findCommunityBySlug();
-
-	if (!community) {
-		return ApiError.COMMUNITY_NOT_FOUND;
-	}
-
-	const authorized = await userCan(
-		Capabilities.editPubType,
-		{ type: MembershipType.community, communityId: community.id },
-		user.id
-	);
-
-	if (!authorized) {
-		return ApiError.UNAUTHORIZED;
-	}
-
-	await db.transaction().execute(async (trx) => {
-		await autoRevalidate(
-			trx.updateTable("_PubFieldToPubType").set({ isTitle: false }).where("B", "=", pubTypeId)
-		).execute();
-		await autoRevalidate(
-			trx
-				.updateTable("_PubFieldToPubType")
-				.set({ isTitle: true })
-				.where("A", "=", pubFieldId)
-				.where("B", "=", pubTypeId)
-		).execute();
-	});
-});
-
-export const removePubField = defineServerAction(async function removePubField(
-	pubTypeId: PubTypesId,
-	pubFieldId: PubFieldsId
-) {
-	const loginData = await getLoginData();
-	if (!loginData || !loginData.user) {
-		return ApiError.NOT_LOGGED_IN;
-	}
-
-	const { user } = loginData;
-
-	const community = await findCommunityBySlug();
-
-	if (!community) {
-		return ApiError.COMMUNITY_NOT_FOUND;
-	}
-
-	const authorized = await userCan(
-		Capabilities.editPubType,
-		{ type: MembershipType.community, communityId: community.id },
-		user.id
-	);
-
-	if (!authorized) {
-		return ApiError.UNAUTHORIZED;
-	}
-
-	await autoRevalidate(
-		db.deleteFrom("_PubFieldToPubType").where("A", "=", pubFieldId).where("B", "=", pubTypeId)
-	).execute();
-});
 
 export const createPubType = defineServerAction(async function createPubType(
 	name: string,
